% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisheryBasics.R
\docType{package}
\name{FisheryBasics}
\alias{FisheryBasics}
\alias{FisheryBasics-package}
\title{Fishery Basic Modelling Package}
\description{
Package containing basics of fishery life cycle modelling using the Beverton-Holt curve. This includes generating data,
a simple model, and a state-based model. All are written using JAGS. All models are based on a harvest rate
and hatchery rate.
}
\author{
Todor Dimitrov \email{todord@uw.edu}
}
