% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genData.R
\name{genData}
\alias{genData}
\title{Generate Sample Data}
\usage{
genData(Nyears, pHOS, HR, processErr, obsErr)
}
\arguments{
\item{Nyears}{Number of years to generate}

\item{pHOS}{Vector contianing proportion of naturally spawning hatchery fish}

\item{HR}{Vector containing harvest rate for every year}

\item{processErr}{Process Error}

\item{obsErr}{Observation Error}
}
\value{
a vector of length Nyears containing the generated data
}
\description{
Generates example data which can be used with the models within the package. If any of Nyears,
pHOS or HR are null then all three take on default values. If processErr is null takes on .1, if
obsErr is null takes on .15.
}
\examples{
N <- 100
pHOS <- sample(rep(seq(0, .9, .1), rep(N/10,10)))
HR <- sample(rep(seq(0, .63, .07), rep(N/10,10)))
processErr <- .1
obsErr <- .15
genData(N, pHOS, HR, processErr, obsErr)

}
