% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleModel.R
\name{simpleModel}
\alias{simpleModel}
\title{Run Simple Model}
\usage{
simpleModel(initValFunc, Sobs, pHOS, HR, Nyears, nIter)
}
\arguments{
\item{initValFunc}{Inital value function for productivity and capacity}

\item{Sobs}{the observed yearly population data}

\item{pHOS}{Vector contianing proportion of naturally spawning hatchery fish}

\item{HR}{Vector containing harvest rate for every year}

\item{Nyears}{Number of years of data}

\item{nIter}{Number of iterations to monitor}
}
\value{
a list containing nIter samples of the posterior distributions of both capacity,
productivity, and Error. Returns NULL if Sobs, pHOS, or HR are not provided or their
lengths do not match.
}
\description{
Runs a basic model to fit a Beverton-Holt curve using the provided observed data, hatchery rates, harvest
rates, and years of data.
}
\examples{
library('rjags')
N <- 100
pHOS <- sample(rep(seq(0, .9, .1), rep(N/10,10)))
HR <- sample(rep(seq(0, .63, .07), rep(N/10,10)))
processErr <- .1
obsErr <- .15
Sobs <- genData(N, pHOS, HR, processErr, obsErr)
initValFunc <- function() {
  list(prod=1, cap=600)
}
simpleModel(initValFunc, Sobs, pHOS, HR, N, 10000)

}
